prepare for cogs - simplify bot.py
